[test] 좌석 예약 가능 여부 단위 테스트 작성
